# Delegate encoding normalization to `vec_normalize_encoding()`