This PR strips out all of the encoding logic from `vec_order_radix()`, replacing it with a single call to `vec_normalize_encoding()` at the top.

This greatly simplifies the character encoding logic, at the cost of some performance in two ways:

- Previously, we only checked the encodings of the _unique_ strings to determine if any re-encoding was required. Now we are forced to loop through the whole vector up front, checking the encoding of each string. This generally isn't that bad, but...

- ...When there are _lots_ of unique values, `vec_normalize_encoding()` is slower than when there are just a few. I assume this has something to do with the CPU caching and reusing the result of `LEVELS(x)` (which we use in `string_is_normalized()`) when there are just a few unique values.

I still think this tradeoff is worth it, as it further standardizes how we handle encodings throughout vctrs and is much easier to reason about. The performance hit from calling `vec_normalize_encoding()` will really only be noticeable if there are a massive amount of unique values.

``` r
library(vctrs)

set.seed(123)

vec_order_radix <- vctrs:::vec_order_radix
vec_normalize_encoding <- vctrs:::vec_normalize_encoding
```

```r
# few uniques
dict <- c("apples", "bananas")
x1 <- sample(dict, size = 1e6, replace = TRUE)
x2 <- sample(dict, size = 1e7, replace = TRUE)

## few uniques / short
# before
bench::mark(vec_order_radix(x1), iterations = 50)
#> # A tibble: 1 x 6
#>   expression               min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_order_radix(x1)   8.72ms   11.3ms      88.6    12.7MB     19.4

# after
bench::mark(vec_order_radix(x1), iterations = 50)
#> # A tibble: 1 x 6
#>   expression               min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_order_radix(x1)   10.6ms   13.3ms      75.4    12.7MB     16.5

# normalization cost
bench::mark(vec_normalize_encoding(x1))
#> # A tibble: 1 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_normalize_encoding(x1)   1.34ms   1.56ms      629.        0B        0


## few uniques / long
# before
bench::mark(vec_order_radix(x2), iterations = 50)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression               min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_order_radix(x2)   98.9ms    104ms      9.45     124MB     9.45

# after
bench::mark(vec_order_radix(x2), iterations = 50)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression               min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_order_radix(x2)    116ms    121ms      8.13     124MB     8.13

# normalization cost
bench::mark(vec_normalize_encoding(x2))
#> # A tibble: 1 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_normalize_encoding(x2)   15.5ms   17.1ms      57.7        0B        0
```

```r
# many uniques
dict <- stringi::stri_rand_strings(200000, length = 6)
y1 <- sample(dict, size = 1e6, replace = TRUE)
y2 <- sample(dict, size = 1e7, replace = TRUE)

## many uniques / short
# before
bench::mark(vec_order_radix(y1), iterations = 50)
#> # A tibble: 1 x 6
#>   expression               min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_order_radix(y1)    103ms    109ms      9.17    43.1MB     6.64

# after
bench::mark(vec_order_radix(y1), iterations = 50)
#> # A tibble: 1 x 6
#>   expression               min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_order_radix(y1)    110ms    120ms      8.34    43.1MB     6.04

# normalization cost
bench::mark(vec_normalize_encoding(y1))
#> # A tibble: 1 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_normalize_encoding(y1)   7.03ms   7.93ms      125.        0B        0


## many uniques / long
# before
bench::mark(vec_order_radix(y2), iterations = 50)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression               min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_order_radix(y2)    452ms    497ms      1.96     258MB     2.16

# after
bench::mark(vec_order_radix(y2), iterations = 50)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression               min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_order_radix(y2)    531ms    570ms      1.71     258MB     1.88

# normalization cost
bench::mark(vec_normalize_encoding(y2))
#> # A tibble: 1 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_normalize_encoding(y2)   79.9ms   87.7ms      11.5        0B        0
```

<sup>Created on 2021-02-26 by the [reprex package](https://reprex.tidyverse.org) (v1.0.0)</sup>