@DavisVaughan 2021-03-10
> It does seem like a big win for code simplicity!

It also makes #1334 much easier to implement

> altrep variants that maintain UTF-8 at all times

Yea that would be great, `vec_normalize_encoding()` could short circuit if it found a `STRSXP` (or whatever is used) that was guaranteed to always be UTF-8